+++ bot-api/dotnet/api/src/internal/BaseBotInternals.cs
@@ -999,6 +999,9 @@ sealed class BaseBotInternals
     {
         _serverHandshake = JsonConverter.FromJson<S.ServerHandshake>(json);
 
+        // Validate bot info before sending bot handshake
+        ValidateBotInfo();
+
         // Reply by sending bot handshake
         var isDroid = _baseBot is Droid;
         var botHandshake = BotHandshakeFactory.Create(_serverHandshake?.SessionId, _botInfo, isDroid, _serverSecret);
@@ -1006,6 +1009,31 @@ sealed class BaseBotInternals
         var text = JsonConverter.ToJson(botHandshake);
 
         _socket.SendTextMessage(text);
+    }
+
+    private void ValidateBotInfo()
+    {
+        if (string.IsNullOrWhiteSpace(_botInfo.Name))
+        {
+            ThrowMissingPropertyException("name");
+        }
+        if (string.IsNullOrWhiteSpace(_botInfo.Version))
+        {
+            ThrowMissingPropertyException("version");
+        }
+        if (_botInfo.Authors.IsNullOrEmptyOrContainsOnlyBlanks())
+        {
+            ThrowMissingPropertyException("authors");
+        }
+    }
+
+    private void ThrowMissingPropertyException(string propertyName)
+    {
+        throw new BotException(
+            $"Required bot property '{propertyName}' is missing. " +
+            "This property is required in order for the bot to be recognized when booting it up and " +
+            "when it needs to join the game. You must set this property in your bot code " +
+            "or provide a .json configuration file.");
     }
 
     private static void VerifyNotNull(Object iEvent, Type eventType)